Fix duplicate instance of Spreadsheet import class
When I revisited this code & prepared for PR I thought I had lost the Spreadsheet class,
not realising I put it in the civiimport - this consolidates the 2 versions into
the extension.

Note that I am on the fence on whether this should be in the extension
or with the others, reflected in the fact I put it in 2 different places on
2 different occasions. I got some support for the extension on chat
so consolidated there
https://chat.civicrm.org/civicrm/pl/xfp4cytepir7jk7axis7aftqye
